requirements_detection.txt
# Requirements for BPMN Detection Demo
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
huggingface_hub>=0.16.0
pytesseract>=0.3.10
wordfreq>=3.0.3
rapidfuzz>=3.6.2
transformers>=4.38.0
sentencepiece>=0.2.0
protobuf>=4.25.0